refactor: prevent whole page of refreshing when searching a new word
